Some fixes for bridges

+++ core/bridges/phpbb3.bridge.class.php
@@ -160,6 +160,9 @@ class phpbb3_bridge extends bridge_generic {
 	}
 	
 	public function autologin($arrCookieData){
+		//If Single Sign On is disabled, abort
+		if ((int)$this->config->get('cmsbridge_disable_sso') == 1) return false;
+		
 		$query = $this->bridgedb->query("SELECT * FROM ".$this->prefix."config");
 		if ($query){
 			while($row = $query->fetchAssoc()){
@@ -238,6 +241,9 @@ class phpbb3_bridge extends bridge_generic {
 	}
 	
 	public function logout() {
+		//If Single Sign On is disabled, abort
+		if ((int)$this->config->get('cmsbridge_disable_sso') == 1) return false;
+		
 		$arrUserdata = $this->bridge->get_userdata($this->user->data['username']);
 		if (isset($arrUserdata['id'])){
 			$this->bridgedb->prepare("DELETE FROM ".$this->prefix."sessions WHERE session_user_id=?")->execute($arrUserdata['id']);
@@ -250,10 +256,10 @@ class phpbb3_bridge extends bridge_generic {
 			}
 		} else return;
 				
-		setcookie($arrConfig['cookie_name'].'_sid', '', 0, $arrConfig['cookie_path'], $arrConfig['cookie_domain'], $arrConfig['cookie_secure']);
+		setcookie($arrConfig['cookie_name'].'_sid', 'somevalue', 0, $arrConfig['cookie_path'], $arrConfig['cookie_domain'], $arrConfig['cookie_secure']);
 		//User-Cookie
-		setcookie($arrConfig['cookie_name'].'_u', '', 0, $arrConfig['cookie_path'], $arrConfig['cookie_domain'], $arrConfig['cookie_secure']);
-		setcookie($arrConfig['cookie_name'].'_k', '', 0, $arrConfig['cookie_path'], $arrConfig['cookie_domain'], $arrConfig['cookie_secure']);
+		setcookie($arrConfig['cookie_name'].'_u', 'somevalue', 0, $arrConfig['cookie_path'], $arrConfig['cookie_domain'], $arrConfig['cookie_secure']);
+		setcookie($arrConfig['cookie_name'].'_k', 'somevalue', 0, $arrConfig['cookie_path'], $arrConfig['cookie_domain'], $arrConfig['cookie_secure']);
 	}
 	
 	public function sync($arrUserdata){
