+++ core/bridges/wbb4.bridge.class.php
@@ -167,7 +167,10 @@ class wbb4_bridge extends bridge_generic {
 	}
 	
 	public function autologin($arrCookieData){
-	$config = array();
+		//If Single Sign On is disabled, abort
+		if ((int)$this->config->get('cmsbridge_disable_sso') == 1) return false;
+		
+		$config = array();
 		$objQuery =  $this->bridgedb->query("SELECT * FROM ".$this->prefix."option WHERE optionName = 'cookie_prefix'");
 		if($objQuery){
 			$result = $objQuery->fetchAllAssoc();
@@ -206,6 +209,9 @@ class wbb4_bridge extends bridge_generic {
 	}
 	
 	public function logout(){
+		//If Single Sign On is disabled, abort
+		if ((int)$this->config->get('cmsbridge_disable_sso') == 1) return false;
+		
 		$arrUserdata = $this->bridge->get_userdata($this->user->data['username']);
 		if (isset($arrUserdata['id'])){
 			$this->bridgedb->prepare("DELETE FROM ".$this->prefix."session WHERE userID=?")->execute($arrUserdata['id']);
@@ -234,9 +240,9 @@ class wbb4_bridge extends bridge_generic {
 		
 		$config['cookie_path'] = (strlen($this->config->get('cmsbridge_sso_cookiepath'))) ? $this->config->get('cmsbridge_sso_cookiepath') : '/';
 		
-		setcookie($config['cookie_prefix'].'cookieHash', '', 0, $config['cookie_path'], $config['cookie_domain'], $this->env->ssl);
-		setcookie($config['cookie_prefix'].'userID', '', 0, $config['cookie_path'], $config['cookie_domain'], $this->env->ssl);
-		setcookie($config['cookie_prefix'].'password', '', 0, $config['cookie_path'], $config['cookie_domain'], $this->env->ssl);
+		setcookie($config['cookie_prefix'].'cookieHash', 'somevalue', 0, $config['cookie_path'], $config['cookie_domain'], $this->env->ssl);
+		setcookie($config['cookie_prefix'].'userID', 'somevalue', 0, $config['cookie_path'], $config['cookie_domain'], $this->env->ssl);
+		setcookie($config['cookie_prefix'].'password', 'somevalue', 0, $config['cookie_path'], $config['cookie_domain'], $this->env->ssl);
 	}
 	
 	private function __checkPassword($username, $password, $hash) {
